Testing: Move scope_and_rse fixture to conftest
